arm: sun4i: prevent memory of disabled features to be reserved
Attached is a patch which prevents memory being reserved for
framebuffer, G2D and video engine if the relevant bits aren't enabled in
the kernel configuration. This whole area looks very much under
construction, and I've no idea whether this is the right approach or
not, but on my server system this gains me an extra 115MB or so of
usable RAM.

(Incidentally, the G2D area is apparently never reserved. Is that correct=
?)